doc-pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doc-pipeline"
version = "0.1.0"
description = "건축·구조 엔지니어링 사내 문서 수집·처리·활용 파이프라인"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "Engineering Team"}
]

dependencies = [
    # PDF text extraction
    "PyMuPDF>=1.24.0",
    # LLM & embedding
    "google-genai>=1.0.0",
    # Vector DB
    "chromadb>=0.5.0",
    # Google Sheets
    "gspread>=6.0.0",
    "google-auth>=2.0.0",
    # File watching
    "watchdog>=4.0.0",
    # Data models
    "pydantic>=2.0.0",
    # Config
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",
    # Multi-format document support
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
]

[project.optional-dependencies]
ocr-marker = [
    "marker-pdf>=0.3.0",
]
ocr-mineru = [
    "magic-pdf>=0.9.0",
]
ocr-gateway = [
    "httpx>=0.27.0",
]
agents = [
    "pydantic-ai>=0.1.0,<2.0",
]
api = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.0",
    "prometheus-fastapi-instrumentator>=7.0",
]
ui = [
    "streamlit>=1.35.0",
]
telegram = [
    "python-telegram-bot>=21.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]
observability = [
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
]
all = [
    "doc-pipeline[api,ocr-marker,ocr-mineru,ocr-gateway,agents,ui,telegram,dev]",
]

[project.scripts]
doc-pipeline = "doc_pipeline.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
